Backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
python-dotenv
pdfplumber
python-docx
google-genai
tree-sitter
tree-sitter-python
openai
langgraph
PyGithub
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-magic; sys_platform == 'darwin'
python-magic-bin; sys_platform == 'win32'
slowapi
httpx
beautifulsoup4
lxml
email-validator
